# Skip logging header for Entry lines with embedded CRs