We don't need a header for each line in a log entry lines with embedded carriage returns.  Print the header once for each entry.

Before
```
BenchmarkFormatNonEmpty-4   	  500000	      2056 ns/op	    4512 B/op	       7 allocs/op
BenchmarkFormatEmpty-4   	 3000000	       494 ns/op	     128 B/op	       3 allocs/op
```

After 
```
BenchmarkFormatNonEmpty-4   	 3000000	       597 ns/op	     224 B/op	       3 allocs/op
BenchmarkFormatEmpty-4      	 3000000	       515 ns/op	     128 B/op	       3 allocs/op
```