@fdawg4l 2017-01-25
> We are modifying the original message from the caller by dropping the trailing new line. If it was the caller's intent to insert a blank line after their message in one call we should let them.

The only place we do this is in `vic-machine`.  Frankly, I don't think we should be using the logging facility to report status to the interactive user.  But, I don't feel like changing it so I can nuke the trailing slash case.

> If we don't want to prefix the header, the code is much simpler; we can just output the message followed by a new line. Personally I think adding the header is much cleaner and easier to read. 

Assuming you're referring to log lines with embedded CRs.  Evaluating every byte in your logfile just to add a header on line breaks is kind of overkill.  I'd rather do without it.  And honestly, if I add a CR to a log message, the header in the preceding line is all I need to debug what happened chronologically.