@hmahmood 2017-01-25
@fdawg4l its pretty simple then: just return `[]byte(t + " " + l + " " + entry.Message + "\n")`.