@hickeng 2017-01-25
@hmahmood @fdawg4l you two are very opinionated about logging formatting, so before any logging changes get made, address the requirements. Please identify:

1. what our primary requirements are for logging, including
  a. when we need to allow multi-line log messages via the logging mechanism (it's a pretty print variety and should therefore be rare or debug only in general)
  b. does logrus support these requirements, if not what are the gaps and how can we address them.
  c. how should trace.Operation work and what does this mean for logging
2. what are our primary requirements for vic-machine user targeted output - using the logging mechanism for this was expedient but a serious abuse as it prevents us from providing status bars for iso upload and any other sort of ncurses style operation

If logrus is the elephant in the room then any changes to logging are pointless until this is addressed