@hmahmood 2017-01-25
@hickeng my initial PR was just to address the logging format, not requirements for logging in general. I intend to take the different issues with logging on one at a time. I generally agree with @fdawg4l on these changes; I just think they can be much simpler for the goals outlined.

Logging format is going to be the same regardless of whether we use logrus or not, so that is why I opted to tackle that first. The only reason in my opinion to use one library over the other would be performance; they all pretty much offer the same features, and we only need a few to be present:

1. support for custom formats
2. support for structured logging, which may help us with tracing

logrus satisfies these right now, so that is why I haven't made any changes to strip it out, and that is why I would like to wait till the other changes are in before worrying about performance.