@hickeng 2017-01-25
@hmahmood where does the requirement for multi-line processing in log messages come from? Is it a pervasive requirement or a rare case where we want pretty print? If it's the latter then using a different logger that's configured for pretty print might make sense instead of scanning all messages.
Without defining a usage requirement for the pretty print case it's hard to identify what the preferred solution is.