@hmahmood 2017-01-25
The multiline requirement mainly comes from debug messages which may include things like a stack trace. The Operation object in particular does that. Also I noticed some variables when printed verbosely have multiline output. I don't think multiline output is the norm for other than debug messages, so you're right it may only be an edge concern when it comes to performance. 

I am not too hung up on how we output the multiline messages, other than we should not modify them. If we take out the `Scanner` in this code, and just output the message without walking it, it would be fine with me.